cleaned code to remove many compiler warnings fixed game config persistence bugs improved asset loading - loading the asset image happens on a background thread now expanded README.md and CONTRIBUTING.md some name refactoring

+++ src/external/cxstructs/cxstructs/SmallVector.h
@@ -90,7 +90,7 @@ class VectorIterator {
 };
 
 namespace cxstructs {
-template <typename T, uint32_t N = 16, typename size_type = uint16_t>
+template <typename T, uint32_t N = 16, typename size_type = int32_t>
 struct SmallVector {
   SmallVector() : size_(0), capacity_(N), data_(stack_data_) {}
